autocomplete: require file path to be specified.

+++ utils/indexing/autocomplete/build_autocomplete.pl
@@ -5,9 +5,11 @@ use warnings;
 
 use FindBin qw($Bin);
 
+die "Please specify path to xml files" unless $ARGV[0];
+
 system("cd $Bin; rm -f autocomplete ; gcc -Ofast -flto -march=native -Wall autocomplete.c -o autocomplete -lm") && die "Cannot compile: $!";
 
-my $files = qx(find . -name \\*.xml);
+my $files = qx(find $ARGV[0] -name \\*.xml);
 
 my $args = "-n 1000000 -c -s all_";
